move sending side code to send.go
